pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weatherlinkv2"
version = "0.1.0"
authors = [
    {name = "Nicolas Mantilla-Molina", email = "nicomantimol1@gmail.com"}
]
description = "Professional Python library for WeatherLink v2 API integration"
readme = "README.md"
license = {text = "CC-BY-NC-SA-4.0"}
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research", 
    "Intended Audience :: Education",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: Creative Commons License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Natural Language :: English",
]
keywords = [
    "weather",
    "meteorology", 
    "weatherlink",
    "api",
    "climate",
    "atmospheric",
    "air quality",
    "environmental monitoring"
]
dependencies = [
    "requests>=2.25.0",
    "pandas>=1.3.0",
    "matplotlib>=3.3.0",
    "seaborn>=0.11.0",
    "python-dotenv>=0.19.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0", 
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.900"
]
viz = [
    "plotly>=5.0",
    "seaborn>=0.11"
]
jupyter = [
    "jupyter>=1.0",
    "ipywidgets>=7.6"
]

[project.urls]
# Homepage = "https://github.com/weatherlink/weatherlinkv2-python"
# Documentation = "https://github.com/weatherlink/weatherlinkv2-python/wiki"
# Repository = "https://github.com/weatherlink/weatherlinkv2-python"
# "Bug Tracker" = "https://github.com/weatherlink/weatherlinkv2-python/issues"

[tool.setuptools]
package-dir = {"" = "."}
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["weatherlinkv2*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
weatherlinkv2 = ["*.txt", "*.md", "*.cfg"]

# Development tools configuration
[tool.black]
line-length = 88
target-version = ["py37"]
include = "\\.pyi?$"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config --tb=short"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests", 
    "unit: marks tests as unit tests",
    "api: marks tests that make API calls"
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning"
]

[tool.coverage.run]
source = ["weatherlinkv2"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:"
]
